Add Neon PostgreSQL support, EF Core migrations, and dev tooling.
MYBUSINESS uses Npgsql/EF6 with a signed EntityFramework6.Npgsql copy under Lib, full dependency set for Npgsql 4.1, placeholder Web.config connection strings, and helper scripts (Apply-NeonToWebConfig, IIS Express, DbMigrator seed). EasyInventory.NeonDb provides EF Core schema and migrator.

Made-with: Cursor

+++ MYBUSINESS/Models/POD.cs
@@ -14,7 +14,7 @@ namespace MYBUSINESS.Models
     
     public partial class POD
     {
-        public decimal Auto { get; set; }
+        public long Auto { get; set; }
         public string POId { get; set; }
         public Nullable<int> PODId { get; set; }
         public Nullable<int> ProductId { get; set; }
